Split application in multipl files and dirs
