+++ server/models/errors.py
@@ -1,0 +1,5 @@
+from sqlmodel import SQLModel
+
+
+class HTTPError(SQLModel):
+    detail: str
